chore(:necktie:): break group source and tests out into individual files
AFFECTS PACKAGES:
@esri/arcgis-rest-groups
