+++ packages/arcgis-rest-groups/src/index.ts
@@ -1,5 +1,11 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-export * from "./groups";
+export * from "./create";
+export * from "./get";
+export * from "./helpers";
 export * from "./notification";
+export * from "./protect";
+export * from "./remove";
+export * from "./search";
+export * from "./update";
