+++ packages/arcgis-rest-groups/test/mocks/responses.ts
@@ -1,11 +1,8 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-import {
-  IGroupSearchResult,
-  IGroupContentResult,
-  IGroupUsersResult
-} from "../../src/groups";
+import { IGroupSearchResult } from "../../src/search";
+import { IGroupContentResult, IGroupUsersResult } from "../../src/get";
 
 import { IGroup } from "@esri/arcgis-rest-common-types";
